IGNITE-25195 Make BinaryReaderHandles package private (#12015)
