+++ modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryUtils.java
@@ -2807,10 +2807,27 @@ public class BinaryUtils {
      * @param ctx Context.
      * @param in Input stream.
      * @param ldr Class loader.
-     * @param hnds Context.
+     * @param reader BinaryReaderEx.
      * @param forUnmarshal {@code True} if reader is need to unmarshal object.
      */
     public static BinaryReaderEx reader(BinaryContext ctx,
+        BinaryInputStream in,
+        ClassLoader ldr,
+        BinaryReaderEx reader,
+        boolean forUnmarshal) {
+        return reader(ctx, in, ldr, reader.handles(), forUnmarshal);
+    }
+
+    /**
+     * Creates reader instance.
+     *
+     * @param ctx Context.
+     * @param in Input stream.
+     * @param ldr Class loader.
+     * @param hnds Context.
+     * @param forUnmarshal {@code True} if reader is need to unmarshal object.
+     */
+    static BinaryReaderEx reader(BinaryContext ctx,
                                         BinaryInputStream in,
                                         ClassLoader ldr,
                                         @Nullable BinaryReaderHandles hnds,
@@ -2824,11 +2841,28 @@ public class BinaryUtils {
      * @param ctx Context.
      * @param in Input stream.
      * @param ldr Class loader.
-     * @param hnds Context.
      * @param skipHdrCheck Whether to skip header check.
      * @param forUnmarshal {@code True} if reader is need to unmarshal object.
      */
     public static BinaryReaderEx reader(BinaryContext ctx,
+        BinaryInputStream in,
+        ClassLoader ldr,
+        boolean skipHdrCheck,
+        boolean forUnmarshal) {
+        return reader(ctx, in, ldr, null, skipHdrCheck, forUnmarshal);
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param ctx Context.
+     * @param in Input stream.
+     * @param ldr Class loader.
+     * @param hnds Context.
+     * @param skipHdrCheck Whether to skip header check.
+     * @param forUnmarshal {@code True} if reader is need to unmarshal object.
+     */
+    static BinaryReaderEx reader(BinaryContext ctx,
                                         BinaryInputStream in,
                                         ClassLoader ldr,
                                         @Nullable BinaryReaderHandles hnds,
