+++ modules/core/src/main/java/org/apache/ignite/internal/client/thin/ClientBinaryMarshaller.java
@@ -22,7 +22,6 @@ import org.apache.ignite.configuration.IgniteConfiguration;
 import org.apache.ignite.internal.binary.BinaryContext;
 import org.apache.ignite.internal.binary.BinaryMarshaller;
 import org.apache.ignite.internal.binary.BinaryMetadataHandler;
-import org.apache.ignite.internal.binary.BinaryReaderHandles;
 import org.apache.ignite.internal.binary.GridBinaryMarshaller;
 import org.apache.ignite.internal.binary.streams.BinaryInputStream;
 import org.apache.ignite.logger.NullLogger;
@@ -65,13 +64,14 @@ class ClientBinaryMarshaller {
     }
 
     /**
-     * Deserializes object from input stream.
+     * Unwrap binary object.
      *
      * @param in Input stream.
-     * @param hnds Object handles.
+     * @param clazz input object class.
+     * @return Binary object.
      */
-    public <T> T deserialize(BinaryInputStream in, BinaryReaderHandles hnds) {
-        return impl.deserialize(in, null, hnds);
+    public <T> T unwrapBinary(BinaryInputStream in, Class<?> clazz) {
+        return impl.unwrapBinary(in, clazz);
     }
 
     /**
