+++ modules/core/src/main/java/org/apache/ignite/internal/jdbc/thin/JdbcThinTcpIo.java
@@ -347,8 +347,7 @@ public class JdbcThinTcpIo {
 
         send(writer.array());
 
-        BinaryReaderEx reader = BinaryUtils.reader(ctx, BinaryStreams.inputStream(read()),
-            null, null, false);
+        BinaryReaderEx reader = BinaryUtils.reader(ctx, BinaryStreams.inputStream(read()), null, false);
 
         boolean accepted = reader.readBoolean();
 
@@ -453,8 +452,7 @@ public class JdbcThinTcpIo {
 
         send(writer.array());
 
-        BinaryReaderEx reader = BinaryUtils.reader(null, BinaryStreams.inputStream(read()),
-            null, null, false);
+        BinaryReaderEx reader = BinaryUtils.reader(null, BinaryStreams.inputStream(read()), null, false);
 
         boolean accepted = reader.readBoolean();
