+++ modules/core/src/main/java/org/apache/ignite/internal/processors/platform/PlatformContextImpl.java
@@ -183,7 +183,6 @@ public class PlatformContextImpl implements PlatformContext, PartitionsExchangeA
         return BinaryUtils.reader(marsh.context(),
             in,
             ctx.config().getClassLoader(),
-            null,
             true,
             true);
     }
